Move StreamStep to another module
